requirements.txt
streamlit>=1.53.0
openai>=1.12.0,<2.0
python-dotenv>=1.0.0
reportlab>=4.4.9
Pillow>=12.1.0
ebooklib>=0.20
requests>=2.31.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
pydub>=0.25.1